perf-collect: Add strict checks for AMD uncore devices
The event encoding format for Data Fabric (amd_df) events has changed in
Zen 4 processors such as Genoa and Bergamo. Even if an older kernel runs
on these processors, it might be programming the counters incorrectly
and this can lead to unexpected results. There is no easy way to detect
this using kernel versions as distro kernels, while using an older base,
may have backports of the upstream patches. Hence, use strict checks for
the event and umask formats that the kernel advertises for the device.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>