perf-collect: Avoid ending groups with cpu-cycles event
On AMD platforms, uncore events will not be used in the following cases.

  * Running the tool inside a guest (VM).
  * Running the tool on bare-metal but without the "amd-uncore" driver.
  * Running the tool on bare-metal with the "amd-uncore" driver but it
    is found that the driver does not have the right level of support
    for the uncore PMUs specific to the processor.

For all these cases, the tool assumes that it is running inside a VM and
applies additional event filtering as seen in commit 9b62dfc ("limit
cycles filter to vm only (not in older kernels)"). This breaks grouping
if "cpu-cycles" is the last event in a group as such groups get merged
with the next group because the entire line, which contains both the
event and the group separator, is ignored.

As a result, some groups are too large, as in, they have more events
than the number of available hardware counters, and hence are invalid,
or as perf tool reports, "<not supported>", as they can never be
scheduled in.

For such groups, redefine the cpu-cycles event as "ls_not_halted_cyc",
its architectural alias, so that it does not get filtered out.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>